Add support for a 'pool' of VM slaves
This makes use of the ComputerListener::preLaunch function to allow us
to configure a maximum number of VMs to allow online at any one time in
a given Lab Manager 'configuration'.  This is useful if your ESX hosts
can only support 5 VMs at once but you have configured 10.  With this
you can set Jenkins to only ever have 5 online at any one time.

Part of this also adds an option to set the boot delay before trying to
connect to the slave.  This must be set large enough such that the VM
will be alive to the point that Jenkins can connect as a slave,
otherwise we mark the bringup attempt as a failure but the VM will be
on.

Signed-off-by: Tom Rini <tom_rini@mentor.com>