LabManagerVirtualMachineSlave: Fix display of non-default launch delay
This was missing a getLaunchDelay function.  While doing this, we
weren't using launchDelay as an int in this class, so just keep it
as a String.  Also, we can add in some form validation here to make
sure we're given a valid value, so do so.  Finally, noticed that we
didn't need a second copy of getVmName, etc, in here.

Signed-off-by: Tom Rini <tom_rini@mentor.com>